+++ src/program/Config.cpp
@@ -139,6 +139,7 @@ void Config::save(const std::string& gamepath) {
     settings.setValue("locale", sc.locale);
     settings.setValue("virtual_steam", sc.virtual_steam);
     settings.setValue("opengl_soft", sc.opengl_soft);
+    settings.setValue("opengl_performance", sc.opengl_performance);
     settings.setValue("async_events", sc.async_events);
     settings.setValue("wait_timeout", sc.wait_timeout);
 
@@ -270,6 +271,7 @@ void Config::load(const std::string& gamepath) {
     sc.savestates_in_ram = settings.value("savestates_in_ram", sc.savestates_in_ram).toBool();
     sc.backtrack_savestate = settings.value("backtrack_savestate", sc.backtrack_savestate).toBool();
     sc.opengl_soft = settings.value("opengl_soft", sc.opengl_soft).toBool();
+    sc.opengl_performance = settings.value("opengl_performance", sc.opengl_performance).toBool();
 
     size = settings.beginReadArray("main_gettimes_threshold");
     for (int t=0; t<size; t++) {